+++ R/PooledConnectionHandler.R
@@ -138,11 +138,21 @@ PooledConnectionHandler <- R6::R6Class(
       self$isActive <- TRUE
     },
 
+    #' Get Connection
+    #' @description
+    #' Returns a connection from the pool
+    #' When the desired frame exits, the connection will be returned to the pool
+    #' @param .deferedFrame  defaults to the parent frame of the calling block.
+    getConnection = function(.deferedFrame = parent.frame()) {
+      conn <- pool::poolCheckout(self$getConnection())
+      withr::defer(pool::poolReturn(conn), envir = .deferedFrame)
+      return(conn)
+    },
+
     #' get dbms
     #' @description Get the dbms type of the connection
     dbms = function() {
-      conn <- pool::poolCheckout(self$getConnection())
-      on.exit(pool::poolReturn(conn))
+      conn <- self$getConnection()
       DatabaseConnector::dbms(conn)
     },
 
@@ -163,8 +173,7 @@ PooledConnectionHandler <- R6::R6Class(
     #' @param sql                                   sql query string
     #' @param snakeCaseToCamelCase                  (Optional) Boolean. return the results columns in camel case (default)
     queryFunction = function(sql, snakeCaseToCamelCase = self$snakeCaseToCamelCase) {
-      conn <- pool::poolCheckout(self$getConnection())
-      on.exit(pool::poolReturn(conn))
+      conn <- self$getConnection()
       data <- DatabaseConnector::dbGetQuery(conn, sql, translate = FALSE)
       if (snakeCaseToCamelCase) {
         colnames(data) <- SqlRender::snakeCaseToCamelCase(colnames(data))
@@ -179,8 +188,7 @@ PooledConnectionHandler <- R6::R6Class(
     #' Overrides ConnectionHandler Call. Does not translate or render sql.
     #' @param sql                                   sql query string
     executeFunction = function(sql) {
-      conn <- pool::poolCheckout(self$getConnection())
-      on.exit(pool::poolReturn(conn))
+      conn <- self$getConnection()
       DatabaseConnector::dbExecute(conn, sql, translate = FALSE)
     }
   )